Change usage to: forked USER/REPO [DIRECTORY]
Now instead of two mandatory arguments (USER and REPOSITORY)
a SINGLE parameter with USER and REPOSITORY delimited with a slash
is required.

It is simpler to copy and paste from GitHub in one go
from the repository page where there is: forked from user/repo

It will also make it possible to make the USER optional in the future
without any problems with keeping the DIRECTORY optional as well.